data packing 2

+++ labs/memory_bound/data_packing/init.cpp
@@ -6,8 +6,8 @@ S create_entry(int first_value, int second_value) {
   S entry;
 
   entry.i = first_value;
-  entry.s = static_cast<short>(second_value);
-  entry.l = static_cast<long long>(first_value * second_value);
+  entry.s = static_cast<uint32_t>(second_value);
+  entry.l = static_cast<uint32_t>(first_value * second_value);
   entry.d = static_cast<float>(first_value) / maxRandom;
   entry.b = first_value < second_value;
